CASMPET-4125: Trim other non-alphanumeric chars from label strings
We ran into problems with a version like `1.4.2-20210427193552+8d374ae` in special cases where the chart name was exactly the right length that the `+` (later translated to an underscore) was the 63rd character.

This fix trims other potential problem characters that might show up in version strings, preventing them being used as labels.

Test:
 - built locally with `helm package`
 - used one of our charts and ran  `helm template . --set global.chart.name=foobarbaz_abcdefghijklmnopqrstuvwxyz12345 --set global.chart.version=1.4.2-20210427193552+8d374ae`
 - this mimicked a string that if truncated to 63 would end with an underscore
 - verified that this truncated it down to `helm.sh/chart: foobarbaz_abcdefghijklmnopqrstuvwxyz12345-1.4.2-20210427193552` which is 62 chars, showing that the `_` is gone.