CASMPET-4921: Bump version to 6.0.0
The old versioning scheme incremented the MINOR version when the first
release was made targeting a new version of CSM. This caused a problem
because consumers of the chart were setting the requirement to pick up
any new chart with the same MAJOR version. This caused a bunch of
charts to pick up CSM-1.1 content when they were rebuilt for CSM-1.0.
The new versioning scheme will change the MAJOR version for new CSM
releases. This commit brings the cray-service chart into alignment with
the new scheme.

The version is 6.0 because this allows the cray-service chart in
CSM-1.1 to be version 5.0.