CASMPET-4920: Bump version to 5.0.0 (1.1)
The old versioning scheme incremented the MINOR version when the first
release was made targeting a new version of CSM. This caused a problem
because consumers of the chart were setting the requirement to pick up
any new chart with the same MAJOR version. This caused a bunch of
charts to pick up CSM-1.1 content when they were rebuilt for CSM-1.0.
The new versioning scheme will change the MAJOR version for new CSM
releases. This commit brings the cray-service chart into alignment with
the new scheme.

The version is 5.0 because there's already a 3.0 & 4.0 for CSM-1.2;
the version in CSM-1.2 is being updated to 6.0.

(cherry picked from commit 27ac47a4bc85d9d59fbeac6ec237cfbab27f1df6)